refactor : projects 분리에 따른 PreviousProjects.tsx 파일 경로 변경
